+++ zsh/.zshrc.d/71-go.zsh
@@ -1,0 +1,5 @@
+# Go Environment
+
+# Define Go system & workspace roots (PATH in 20-paths.zsh)
+export GOROOT="/usr/local/go"
+export GOPATH="$HOME/go"