+++ zsh/.zshrc.d/73-java.zsh
@@ -1,0 +1,4 @@
+# Java Environment
+
+# Set Java home path (PATH in 20-paths.zsh)
+export JAVA_HOME="/usr/lib/jvm/java-21-openjdk"